all: Rename storeutil pkg to store and use alias
